# Conflicts:
#	business/member-api/src/main/java/com/lawu/eshop/member/api/controller/DiscountPackageController.java
#	business/merchant-api/src/main/java/com/lawu/eshop/merchant/api/controller/AdController.java
#	dto/property-dto/src/main/java/com/lawu/eshop/property/dto/BankAccountDTO.java
#	service-share/ad-srv/src/main/java/com/lawu/eshop/ad/srv/controller/AdController.java
#	service-share/ad-srv/src/main/java/com/lawu/eshop/ad/srv/converter/AdConverter.java
#	service-share/ad-srv/src/main/java/com/lawu/eshop/ad/srv/service/impl/AdServiceImpl.java
#	service-share/mall-srv/src/main/java/com/lawu/eshop/mall/srv/service/impl/CommentMerchantServiceImpl.java
#	service-share/mall-srv/src/main/java/com/lawu/eshop/mall/srv/service/impl/CommentProductServiceImpl.java
#	service-share/property-srv/src/main/java/com/lawu/eshop/property/srv/bo/BankAccountBO.java
#	service-share/user-srv/src/main/java/com/lawu/eshop/user/srv/service/impl/FavoriteMerchantServiceImpl.java
#	service-share/user-srv/src/main/java/com/lawu/eshop/user/srv/service/impl/MemberServiceImpl.java

+++ dto/property-dto/src/main/java/com/lawu/eshop/property/dto/BankAccountDTO.java
@@ -21,9 +21,9 @@ public class BankAccountDTO {
 
 	@ApiModelProperty(value = "银行卡图标", required = true)
 	private String iconUrl;
-	
+
 	@ApiModelProperty(value = "所属银行主键", required = true)
-	private Long bankId;
+	private Integer bankId;
 
 	public Long getId() {
 		return id;
@@ -73,13 +73,13 @@ public class BankAccountDTO {
 		this.iconUrl = iconUrl;
 	}
 
-	public Long getBankId() {
+	public Integer getBankId() {
 		return bankId;
 	}
 
-	public void setBankId(Long bankId) {
+	public void setBankId(Integer bankId) {
 		this.bankId = bankId;
 	}
-	 
-	
+
+
 }
